+++ Resolution/IInstanceResolver.cs
@@ -3,7 +3,7 @@ using System;
 namespace Void2610.LiminalPalette
 {
     /// <summary>
-    /// インスタンスメソッドの [ConsoleCommand] を実行する際に、
+    /// インスタンスメソッドの [LiminalCommand] を実行する際に、
     /// メソッドが属する型のインスタンスを解決するための抽象。
     /// VContainer / Zenject など DI コンテナと連携する想定。
     /// 解決できない場合は null を返す (例外は投げない)。